requirements.txt
sentence-transformers
chromadb
google-genai
python-dotenv
numpy
torch
